reduce memory usage of Text_token (the "the +100000 is used to prevent the addition" is no needed because string_hash can grown)

+++ korean/String_hash2.h
@@ -1188,7 +1188,7 @@ public:
 	
 		while((*scan)){
 			if( addr >= (*scan)->pageAddr){
-				offset = addr - (*scan)->pageAddr;
+				offset = (int)(addr - (*scan)->pageAddr);
 				if(offset < pgSize){
 					return( (*scan)->pageNum * pgEMcnt+
 						offset/pgESize);
